Add pekko-http-circe-base
Adds a base module to include Circe-specific support, but without
exposing the actual parser implementation.
